requirements.txt
tqdm==4.65.0
Pillow
opencv-python
imageio
matplotlib
kornia==0.7.3
natsort
pyyaml
wandb
lpips
open3d==0.16.0
torchmetrics==1.5.0
cyclonedds
pytorch-msssim
plyfile==0.8.1
einops
evo
trimesh
git+https://github.com/JonathonLuiten/diff-gaussian-rasterization-w-depth.git@cb65e4b86bc3bd8ed42174b72a62e8d3a3a71110
git+https://github.com/eriksandstroem/evaluate_3d_reconstruction_lib.git@9b3cc08be5440db9c375cc21e3bd65bb4a337db7

submodules/Hierarchical-Localization 
submodules/simple-knn 
submodules/gaus_2dgs_rasterization
